+++ packages/vscode-extension/CHANGELOG.md
@@ -4,7 +4,7 @@
 Incremental version for Teams Toolkit with multiple bugs fixes and the following updates:
 
 New Features:
-- User can use Teams Toolkit to create CI/CD pipelines for GitHub, Azure DevOps and Jenkins.
+- User can use Teams Toolkit to create workflow automation templates for Github, Azure DevOps and Jenkins.
 
 Enhancement:
 - Enhance TeamsFx SDK.